Add shared event reporting crate
