+++ crates/agent-lite/Cargo.toml
@@ -12,8 +12,8 @@ readme = "../../README.md"
 [dependencies]
 aya = { version = "0.13" }
 bpf-api = { package = "qqrm-bpf-api", version = "0.1.0", path = "../bpf-api" }
-serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+event-reporting = { version = "0.1.0", path = "../event-reporting" }
 
 anyhow = "1"
 log = "0.4"
