+++ crates/cli/Cargo.toml
@@ -22,6 +22,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 toml = "0.8"
 qqrm-agent-lite = { version = "0.1.0", path = "../agent-lite" }
+event-reporting = { version = "0.1.0", path = "../event-reporting" }
 
 [dev-dependencies]
 serial_test = "3"
